+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #11 Upgrade to Bootique 0.13
 * #12 Support for init parameters for servlets and filters
 * #15 Support for sessions in Jetty
+* #19 Support for context init parameters
 
 ## 0.12: